+++ docs/RELEASE-NOTES.rst
@@ -26,6 +26,7 @@ Added Functionality
 
 Bug Fixes
 `````````
+* `Issue 2726 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/2726>`_: Fix prometheus metrics broken in v2.11.1
 
 Vulnerability Fixes
 ```````````````````